+++ packages/subgraph/manifest/subgraph.placeholder.yaml
@@ -142,7 +142,7 @@ templates:
           handler: handleSignatureValidatorSet
         - event: NewURI(string)
           handler: handleNewURI
-  - name: DaoTemplateV1_2_0
+  - name: DaoTemplateV1_3_0
     kind: ethereum/contract
     network: {{network}}
     source:
@@ -151,7 +151,7 @@ templates:
       kind: ethereum/events
       apiVersion: 0.0.5
       language: wasm/assemblyscript
-      file: ./src/dao/dao_v1_2_0.ts
+      file: ./src/dao/dao_v1_3_0.ts
       entities:
         - Dao
       abis:
